Explicit cast to make VS quiet (?)

+++ NewKernel_d/include/CGAL/NewKernel_d/function_objects_cartesian.h
@@ -681,7 +681,7 @@ template <class R_> struct Construct_circumcenter : Store_kernel<R_> {
       typedef typename LAd::Square_matrix Matrix;
       typedef typename LAd::Vector Vec;
       typename Get_functor<R_, Scalar_product_tag>::type sp(this->kernel());
-      int k=std::distance(f,e);
+      int k=static_cast<int>(std::distance(f,e));
       int d=pd(p0);
       Matrix m(k,k);
       Vec b(k);
